be a bit more conservative about deleting empty dirs as it can make debugging tricky after failures if the empty dirs were improtant; this is a good compromise for now

+++ pysys/baserunner.py
@@ -284,8 +284,10 @@ class BaseRunner(ProcessUser):
 								time.sleep(0.1)
 								count = count + 1
 								
-				# always try to delete empty directories (just as we do for empty files)
-				if deleted == len(filenames):
+				# always try to delete empty directories (just as we do for empty files); 
+				# until we have some kind of internal option for disabling this for debugging 
+				# purpose only delete dirs when we've just deleted the contents ourselves 
+				if deleted > 0 and deleted == len(filenames):
 					try:
 						os.rmdir(dirpath)
 					except Exception as ex:
